move inline code from renf.h to proper implementation files
