+++ e-antic/renf.h
@@ -50,23 +50,9 @@ void renf_refine_embedding(renf_t nf, slong prec);
 
 int renf_set_immutable(renf_t nf, int immutable);
 
-static __inline__
-int nf_equal(const nf_t nf1, const nf_t nf2)
-{
-    return fmpq_poly_equal(nf1->pol, nf2->pol);
-}
+int renf_equal(const renf_t nf1, const renf_t nf2);
 
-static __inline__
-int renf_equal(const renf_t nf1, const renf_t nf2)
-{
-    return nf_equal(nf1->nf, nf2->nf) && arb_overlaps(nf1->emb, nf2->emb);
-}
-
-static __inline__
-slong renf_degree(renf_t nf)
-{
-    return fmpq_poly_degree(nf->nf->pol);
-}
+slong renf_degree(renf_t nf);
 
 /* Randomisation */
